+++ crates/polars-core/src/frame/mod.rs
@@ -1,7 +1,6 @@
 //! DataFrame module.
 #[cfg(feature = "zip_with")]
 use std::borrow::Cow;
-use std::cmp::min;
 use std::{mem, ops};
 
 use polars_utils::itertools::Itertools;
@@ -27,7 +26,6 @@ pub mod row;
 mod top_k;
 mod upstream_traits;
 
-use arrow::bitmap::MutableBitmap;
 use arrow::record_batch::RecordBatch;
 use polars_utils::pl_str::PlSmallStr;
 #[cfg(feature = "serde")]
@@ -3003,126 +3001,6 @@ impl DataFrame {
         }
         DataFrame::new(new_cols)
     }
-
-    /// Inequality join. Matches rows from this DataFrame with rows from another DataFrame
-    /// using two inequality operators (one of [<, <=, >, >=]).
-    /// Based on Khayyat et al. 2015, "Lightning Fast and Space Efficient Inequality Joins"
-    /// and extended to work with duplicate values.
-    #[allow(clippy::too_many_arguments)]
-    pub fn ie_join(
-        &self,
-        other: &DataFrame,
-        col1: &str,
-        op1: &str,
-        other_col1: &str,
-        col2: &str,
-        op2: &str,
-        other_col2: &str,
-    ) -> PolarsResult<Self> {
-        let mut x = self.column(col1)?.to_physical_repr().into_owned();
-        x.extend(&other.column(other_col1)?.to_physical_repr())?;
-
-        let mut y = self.column(col2)?.to_physical_repr().into_owned();
-        y.extend(&other.column(other_col2)?.to_physical_repr())?;
-
-        // Determine the sort order based on the comparison operators used.
-        // We want to sort L1 so that "x[i] op1 x[j]" is true for j > i,
-        // and L2 so that "y[i] op2 y[j]" is true for j < i
-        // (except in the case of duplicates and strict inequalities).
-        // Note that the algorithms published in Khayyat et al. have incorrect logic for
-        // determining whether to sort descending.
-        let l1_descending = op1 == ">" || op1 == ">=";
-        let l2_descending = op2 == "<" || op2 == "<=";
-
-        let l1_sort_options = SortOptions::default()
-            .with_maintain_order(true)
-            .with_nulls_last(false)
-            .with_order_descending(l1_descending);
-        // Get ordering of x, skipping any null entries as these cannot be matches
-        let l1_order = x
-            .arg_sort(l1_sort_options)
-            .slice(x.null_count() as i64, x.len() - x.null_count());
-
-        let l1_array = with_match_physical_numeric_polars_type!(x.dtype(), |$T| {
-            let ca: &ChunkedArray<$T> = x.as_ref().as_ref().as_ref();
-            build_l1_array(ca, &l1_order, self.height() as IdxSize)
-        })?;
-
-        let y_ordered = y.take(&l1_order)?;
-        let l2_sort_options = SortOptions::default()
-            .with_maintain_order(true)
-            .with_nulls_last(false)
-            .with_order_descending(l2_descending);
-        // Get the indexes into l1, ordered by y values.
-        // l2_order is the same as "p" from Khayyat et al.
-        let l2_order = y_ordered.arg_sort(l2_sort_options).slice(
-            y_ordered.null_count() as i64,
-            y_ordered.len() - y_ordered.null_count(),
-        );
-
-        // Create a bit array with order corresponding to L1,
-        // denoting which entries have been visited while traversing L2.
-        let mut bit_array = FilteredBitArray::from_len_zeroed(l1_order.len());
-
-        let mut left_row_ids = PrimitiveChunkedBuilder::<IdxType>::new("left_indices", 0);
-        let mut right_row_ids = PrimitiveChunkedBuilder::<IdxType>::new("right_indices", 0);
-
-        if op2 == ">" || op2 == "<" {
-            // For strict inequalities, we rely on using a stable sort of l2 so that
-            // p values only increase as we traverse a run of equal y values.
-            // To handle inclusive comparisons in x and duplicate x values we also need the
-            // sort of l1 to be stable, so that the left hand side entries come before the right
-            // hand side entries (as we mark visited entries from the right hand side).
-            for p in l2_order.into_no_null_iter() {
-                l1_array.process_entry(
-                    p as usize,
-                    &mut bit_array,
-                    op1,
-                    &mut left_row_ids,
-                    &mut right_row_ids,
-                );
-            }
-        } else {
-            let l2_array = with_match_physical_numeric_polars_type!(y.dtype(), |$T| {
-                let ca: &ChunkedArray<$T> = y_ordered.as_ref().as_ref().as_ref();
-                build_l2_array(ca, &l2_order)
-            })?;
-            // For non-strict inequalities in l2, we need to track runs of equal y values and only
-            // check for matches after we reach the end of the run and have marked all rhs entries
-            // in the run as visited.
-            let mut run_start = 0;
-            for i in 0..l2_array.len() {
-                let p = l2_array[i].l1_index;
-                l1_array.mark_visited(p as usize, &mut bit_array);
-                if l2_array[i].run_end {
-                    for l2_item in l2_array.iter().take(i + 1).skip(run_start) {
-                        let p = l2_item.l1_index;
-                        l1_array.process_lhs_entry(
-                            p as usize,
-                            &bit_array,
-                            op1,
-                            &mut left_row_ids,
-                            &mut right_row_ids,
-                        );
-                    }
-                    run_start = i + 1;
-                }
-            }
-        }
-
-        let left_rows = self.take(&left_row_ids.finish())?;
-        let mut right_rows = other.take(&right_row_ids.finish())?;
-
-        let left_col_names = PlHashSet::from_iter(left_rows.get_column_names_owned());
-        let right_col_names = right_rows.get_column_names_owned();
-        for column_name in right_col_names {
-            if left_col_names.contains(&column_name) {
-                let new_name = column_name.clone() + "_right";
-                right_rows.rename(&column_name, &new_name)?;
-            }
-        }
-        left_rows.hstack(&right_rows.columns)
-    }
 }
 
 pub struct RecordBatchIter<'a> {
@@ -3209,266 +3087,6 @@ fn ensure_can_extend(left: &Series, right: &Series) -> PolarsResult<()> {
         left.name(), right.name(),
     );
     Ok(())
-}
-
-/// Item in L1 array used in the IEJoin algorithm
-#[derive(Clone, Copy, Debug)]
-struct L1Item<T> {
-    /// 1 based index for entries from the LHS df, or -1 based index for entries from the RHS
-    row_index: i64,
-    /// X value
-    value: T,
-}
-
-/// Item in L2 array used in the IEJoin algorithm
-#[derive(Clone, Copy, Debug)]
-struct L2Item {
-    /// Corresponding index into the L1 array of
-    l1_index: IdxSize,
-    /// Whether this is the end of a run of equal y values
-    run_end: bool,
-}
-
-trait L1Array {
-    fn process_entry(
-        &self,
-        l1_index: usize,
-        bit_array: &mut FilteredBitArray,
-        op1: &str,
-        left_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-        right_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-    );
-
-    fn process_lhs_entry(
-        &self,
-        l1_index: usize,
-        bit_array: &FilteredBitArray,
-        op1: &str,
-        left_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-        right_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-    );
-
-    fn mark_visited(&self, index: usize, bit_array: &mut FilteredBitArray);
-}
-
-fn find_search_start_index<T>(l1_array: &[L1Item<T>], index: usize, operator: &str) -> usize
-where
-    T: NumericNative,
-    T: TotalOrd,
-{
-    let value = l1_array[index].value;
-
-    if operator == ">" || operator == "<" {
-        // Search forward until we find a value not equal to the current x value
-        let mut left_bound = index;
-        let mut right_bound = index + 1;
-        let mut step_size = 1;
-        while right_bound < l1_array.len() && l1_array[right_bound].value.tot_eq(&value) {
-            left_bound = right_bound;
-            right_bound = min(right_bound + step_size, l1_array.len());
-            step_size *= 2;
-        }
-        // Now binary search to find the first value not equal to the current x value
-        while right_bound - left_bound > 1 {
-            let mid = left_bound + (right_bound - left_bound) / 2;
-            if l1_array[mid].value.tot_eq(&value) {
-                left_bound = mid;
-            } else {
-                right_bound = mid;
-            }
-        }
-        right_bound
-    } else {
-        // Search backwards to find the first value equal to the current x value
-        let mut left_bound = if index > 0 { index - 1 } else { 0 };
-        let mut right_bound = index;
-        let mut step_size = 1;
-        while left_bound > 0 && l1_array[left_bound].value.tot_eq(&value) {
-            right_bound = left_bound;
-            left_bound = if left_bound > step_size {
-                left_bound - step_size
-            } else {
-                0
-            };
-            step_size *= 2
-        }
-
-        if l1_array[left_bound].value.tot_eq(&value) {
-            return left_bound;
-        }
-
-        // Now binary search to find the first value equal to the current x value
-        while right_bound - left_bound > 1 {
-            let mid = left_bound + (right_bound - left_bound) / 2;
-            if l1_array[mid].value.tot_eq(&value) {
-                right_bound = mid;
-            } else {
-                left_bound = mid;
-            }
-        }
-        right_bound
-    }
-}
-
-impl<T> L1Array for Vec<L1Item<T>>
-where
-    T: NumericNative,
-{
-    fn process_entry(
-        &self,
-        l1_index: usize,
-        bit_array: &mut FilteredBitArray,
-        op1: &str,
-        left_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-        right_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-    ) {
-        let row_index = self[l1_index].row_index;
-        let from_lhs = row_index > 0;
-        if from_lhs {
-            // This entry comes from the left hand side DataFrame.
-            // Find all following entries in L1 (meaning they satisfy the first operator)
-            // that have already been visited (so satisfy the second operator).
-            // Because we use a stable sort for l2, we know that we won't find any
-            // matches for duplicate y values when traversing forwards in l1.
-            let start_index = find_search_start_index(self, l1_index, op1);
-            bit_array.on_set_bits_from(start_index, |set_bit: usize| {
-                let right_row_index = self[set_bit].row_index;
-                debug_assert!(right_row_index < 0);
-                left_row_ids.append_value((row_index - 1) as IdxSize);
-                right_row_ids.append_value((-right_row_index) as IdxSize - 1);
-            });
-        } else {
-            bit_array.set_bit(l1_index);
-        }
-    }
-
-    fn process_lhs_entry(
-        &self,
-        l1_index: usize,
-        bit_array: &FilteredBitArray,
-        op1: &str,
-        left_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-        right_row_ids: &mut PrimitiveChunkedBuilder<IdxType>,
-    ) {
-        let row_index = self[l1_index].row_index;
-        let from_lhs = row_index > 0;
-        if from_lhs {
-            let start_index = find_search_start_index(self, l1_index, op1);
-            bit_array.on_set_bits_from(start_index, |set_bit: usize| {
-                let right_row_index = self[set_bit].row_index;
-                debug_assert!(right_row_index < 0);
-                left_row_ids.append_value((row_index - 1) as IdxSize);
-                right_row_ids.append_value((-right_row_index) as IdxSize - 1);
-            });
-        }
-    }
-
-    fn mark_visited(&self, index: usize, bit_array: &mut FilteredBitArray) {
-        let from_lhs = self[index].row_index > 0;
-        if !from_lhs {
-            bit_array.set_bit(index);
-        }
-    }
-}
-
-fn build_l1_array<T>(
-    ca: &ChunkedArray<T>,
-    order: &IdxCa,
-    right_df_offset: IdxSize,
-) -> PolarsResult<Box<dyn L1Array>>
-where
-    T: PolarsNumericType,
-{
-    let mut array: Vec<L1Item<T::Native>> = Vec::with_capacity(ca.len());
-    for index in order.into_no_null_iter() {
-        let value = ca
-            .get(index as usize)
-            // Nulls should have been skipped over
-            .ok_or_else(|| polars_err!(ComputeError: "Unexpected null value in IEJoin data"))?;
-        let row_index = if index < right_df_offset {
-            index as i64 + 1
-        } else {
-            -((index - right_df_offset) as i64) - 1
-        };
-        array.push(L1Item { row_index, value });
-    }
-    Ok(Box::new(array))
-}
-
-fn build_l2_array<T>(ca: &ChunkedArray<T>, order: &IdxCa) -> PolarsResult<Vec<L2Item>>
-where
-    T: PolarsNumericType,
-    T::Native: TotalOrd,
-{
-    let mut array = Vec::with_capacity(ca.len());
-    let mut prev_index = 0;
-    let mut prev_value = T::Native::default();
-    for (i, l1_index) in order.into_no_null_iter().enumerate() {
-        let value = ca
-            .get(l1_index as usize)
-            // Nulls should have been skipped over
-            .ok_or_else(|| polars_err!(ComputeError: "Unexpected null value in IEJoin data"))?;
-        if i > 0 {
-            array.push(L2Item {
-                l1_index: prev_index,
-                run_end: value.tot_ne(&prev_value),
-            });
-        }
-        prev_index = l1_index;
-        prev_value = value;
-    }
-    if !order.is_empty() {
-        array.push(L2Item {
-            l1_index: prev_index,
-            run_end: true,
-        });
-    }
-    Ok(array)
-}
-
-/// Bit array with a filter to speed up searching for set bits,
-/// based on section 4.1 in Khayyat et al.
-struct FilteredBitArray {
-    bit_array: MutableBitmap,
-    filter: MutableBitmap,
-}
-
-impl FilteredBitArray {
-    const CHUNK_SIZE: usize = 1024;
-
-    pub fn from_len_zeroed(len: usize) -> Self {
-        Self {
-            bit_array: MutableBitmap::from_len_zeroed(len),
-            filter: MutableBitmap::from_len_zeroed(len.div_ceil(Self::CHUNK_SIZE)),
-        }
-    }
-
-    pub fn set_bit(&mut self, index: usize) {
-        self.bit_array.set(index, true);
-        self.filter.set(index / Self::CHUNK_SIZE, true);
-    }
-
-    pub fn on_set_bits_from<F>(&self, start: usize, mut action: F)
-    where
-        F: FnMut(usize),
-    {
-        let start_chunk = start / Self::CHUNK_SIZE;
-        let mut chunk_offset = start % Self::CHUNK_SIZE;
-        for chunk_idx in start_chunk..self.filter.len() {
-            if self.filter.get(chunk_idx) {
-                // There are some set bits in this chunk
-                let start = chunk_idx * Self::CHUNK_SIZE + chunk_offset;
-                let end = min((chunk_idx + 1) * Self::CHUNK_SIZE, self.bit_array.len());
-                for bit_idx in start..end {
-                    // SAFETY: `bit_idx` is always less than `self.bit_array.len()`
-                    if unsafe { self.bit_array.get_unchecked(bit_idx) } {
-                        action(bit_idx);
-                    }
-                }
-            }
-            chunk_offset = 0;
-        }
-    }
 }
 
 #[cfg(test)]
